[homework-7] First attempt

+++ library_data_jpa/src/main/java/ru/otus/homework7/service/AuthorService.java
@@ -4,6 +4,6 @@ import ru.otus.homework7.domain.Author;
 
 public interface AuthorService {
 
-    Author getAuthorById(int authorId);
+    Author getAuthorById(long authorId);
     Author getAuthorByName(String authorName);
 }
