+++ library_data_jpa/src/main/java/ru/otus/homework7/service/CommentsService.java
@@ -5,9 +5,9 @@ import ru.otus.homework7.domain.Comment;
 import java.util.List;
 
 public interface CommentsService {
-    Comment getCommentById(int commentId);
-    List<Comment> getCommentsByBookId(int bookId);
-    void saveNewComment(int bookId);
-    void updateComment(int commentId, String comment);
-    void deleteComment(int commentId);
+    Comment getCommentById(long commentId);
+    List<Comment> getCommentsByBookId(long bookId);
+    void saveNewComment(long bookId);
+    void updateComment(long commentId, String comment);
+    void deleteComment(long commentId);
 }
